Updated Firebase support
