+++ syncmanager/app/src/main/res/values/strings.xml
@@ -39,6 +39,8 @@
     <string name="channel_not_exists">Channel doesn\'t exists</string>
     <string name="no_members">No Members</string>
     <string name="blocked">Blocked</string>
+	<string name="message_received">Somebody sent you a message.</string>
+    <string name="register_push_token">Connection required to register push token.</string>
 
     <!-- main -->
     <string name="main_drawer_open">Open navigation drawer</string>
